Fix #5, Update rwg-charter.md with dates
Added better dates, imagining a July 1 announce and an election in August (after BSidesLV/Black Hat/Def Con). Also updated some language from just CNAs to CNAs and the CVE Board, and an imagined URL to the Groups.IO group.

This is still a pre-release version 0.2, and hasn't been adopted yet. That'll likely happen in August as well, once we let people know about the working group and how to join.